[OO] Optimized Parrot_pmc_get_type_str() slightly.
Because the classname -> integer mapper is an internal implementation detail
which should never leak out of the OO subsystem, it's okay to break the VTABLE
encapsulation in this case. This provides a 2.33% performance improvement on
the OO-heavy stress1.pasm benchmark and should also help any other code which
creates objects based on string class names.

Arguably Parrot should resolve string literals to class objects at compilation
or optimization time, but that's a larger change.

As this commit includes a Makefile dependency addition, reconfigure
recommended, but not required.