POST::PBCCompler - Add get_or_create_*
The PackfileConstantTable apparently had different functions for
different types.  Now it just has one name and acts based on the type
of thing.  Since NQP stores everything as a PMC we need different
functions.  The find_caller_lex bit may be a bit strange, but it makes
the conversion faster.