Push clone_key_arg() closer to where it is needed.
To simplify *_keyed() vtable overrides, key arguments to these need to be cloned
with any register references copied into the clone. However, such overrides are
rare and checking this for EVERY SINGLE PMC ARGUMENT EVER is serious overkill on
a hotpath.

Instead, we need to clone the potential keys only in the code that sets up the
keyed overrides. This gives a modest speedup on PCC-heavy code (~2% on oofib).