Deduplicate m0 C tests by adding a M0_INTERP env var
We now run the C implementation with exactly the same test files as the
Perl implementation.  The m0_c_tests Makefile target merely sets a
M0_INTERP env var to tell the test which binary it should be testing
against.

In the future, we will want something to specify the m0 assembler to
use, but that is not an issue now.

Currently, the m0_c_tests target has some expected failures. This should
motivate us to make them pass, or come up with a "fudge framework."