[pct]:  Switch HLLCompiler's .lineof to a binary search instead of linear.
Recent profiling from mls++ on Rakudo's setting compilation seems to
indicate that lineof does a lot of work.  This patch switches the
linear search to a binary search, making an O(n**2) process into
an O(n*log(n)) one.  However, this doesn't seem to translate into
any sort of significant speed improvement in setting compilation,
which makes me think the profiling itself is off.   Still, it's an
easy optimization for now so I'm going ahead and committing it.