Add get_string to Fixed Float and Integer Array
Also adds tests for get_string to Boolean and PMC.

This makes most Fixed*Array consistent.  The exception is
FixedPMCArray which prints the number of elements instead of the
elements themselves.  We may wish to change that, but I simply wanted
to add the missing strings to avoid exceptions when printing.