Fix sleep in threads
Only up to 4 threads are now created. These threads are stored in a
threads array and get notified on alarms by the main thread writing a
byte to the notifier pipe of the thread.

TODO: implement on Windows
TODO: only write to the pipe if the thread is actually waiting for
alarms