Fix scheduling updates on data owning interpreters
Storing references to foreign interpreters and probably other objects as
well screws up garbage collection and leads to all kinds of strange
errors.

Accessing attributes or methods of the proxy pmc itself leads to a whole
lot of confusion about for example which namespace to search.

Working solution is a new schedule_proxied method which keeps access to
the proxy's attributes and the foreign interp in C space where it's
safe.