new autoclose implementation, no longer relying on sub->ctx
sub->ctx contained the context of the last call to this sub. It
was used in the autoclose code, which does not work when
multiple threads are running. Thus we do away with sub->ctx, but
search the call chain for a suitable context instead.
We also delete the now unused SUB_FLAG_IS_OUTER flag.