disable auto-finalize for now
To finalize, we need to get the exception with get_results.
Unfortunately, get_results clears the caller object, so other
code also doing a get_results to get the exception will no
longer work. I don't know how to fix that except to add a
new op that just "peeks" at the exception.