Make t/pmc/nci.t more reliable
One of the tests failed sometimes on some machines. It tests asynchronous
callbacks which get scheduled as tasks. The sleep 0.001 gave the scheduler
a chance to stop the current task and work through all the posted async
callbacks. But one millisecond might not be enough time on a slow or loaded
system to get all callbacks done. On expiry of the alarm, the interrupted
Task gets scheduled for immediate execution again, so the next test could
start before all callbacks are done.

Fixed it by calling pass in a loop until the callbacks are all done.