Fix recursion_depth growing with every task started
Restore the recursion_depth in the current context after invoking a
task's sub. This is a workaround for Parrot_Sub_invoke incrementing
recursion_depth unconditionally which is not true for invoked tasks,
since they run in parallel to the invoking sub, not nested but still
inherit the context.

This patch also unifies recursion_depth accessors to return the new
value, fixing an off by one error in recursion depth limiting.