Distribute scheduled tasks onto worker threads
The scheduler now tries to be a little smarter by finding the worker
thread with the lowest task count and scheduling the new task onto this
thread.

The downside is that scheduling a task needs to allocate new PMCs in the
thread interp. To not run into concurrency problems in the GC, the
scheduler locks a mutex while running a task, so the scheduling may have
to wait until the current task is preempted.