Proxy namespaces and classes as well
With this change, threads have read only access to their parent's
namespaces and classes. Though this avoids any problems with cloning
namespaces and allows task tests to use the imported ok function and
thread tests to pass, it's still not perfect.

Classes can be instantiated on a thread if they have been instantiated
before in the main thread.