[ci] Turn on IRC notifications
Our test suite (smolder_fulltest) is now passing on Travis CI [0] .
Every commit is being tested with clang, gcc and g++ in a disposable
Ubuntu VM container.

We really need "make allhlltest", but that is so long that it would
timeout with their current settings.

Next step is Rakudo on Travis.

[0] http://travis-ci.org/#!/parrot/parrot/builds/659732