Create the beginning of an embedded Parrot "breed"
I was previously calling this a "flavor", but particle++ suggested breed
and so it shall be.

This embedded breed leaves out everything except the very core
functionality of Parrot and hence is best suited to small resource
embedded systems. In particular, it does not include GMP, ICU, OpenGL,
gettext or any other convenience library.