Don't create proxies for Subs but clone them
When executing a proxied Sub, the context belonging to the foreign
interp would be used. This context for example contains globals and
namespaces.
get_global in this Sub would therefore fetch a global from an unproxied
namespace and get an unproxied global in return without any chance for
us to catch this and create a proxy.

Better to clone the sub and create proxies for all its PMC attributes
(most of all namespaces)