Keep the GC blocked until the proxy for the result is created.
We were blocking the GC while the call on the proxied PMC was
running to prevent it from finding foreign PMCs on the stack. But even
after the call, there's still the result! When creating the proxy for
the result, the GC was already unblocked and every 100,000th proxied
call or so it would start marking at this exact moment and find the
result PMC.

Fixed by simply unblocking the GC only after the proxy is created and
the original being no longer on the stack.