Disable task_primes.t for now - too much NYI
In task_primes.t tasks create other tasks and communicate heavily with
send/receive. The current threading implementation does not yet support
such a use case. It's limited to having a master task create the child
tasks which only communicate with their master. However, this is purely an
implementation detail and not an architectural restriction.