Remove bad test from t/dynoplibs/io.t.
Overwriting a class by name in the global class hash IS NOT how we provide the ability to use custom classes in place of native ones. The mechanism to use is .hll_map. Relying on the behavior (or even continued existance) of the global class hash is a bad idea in all user-facing cases. kid51++ for reporting this failure.