First writeup of the proposed new IO subsystem architecture.
Operations for specific IO types (file, socket, pipe, etc) are broken into separate vtables for easy non-switch dispatch. Buffering will be encapsulated as a separate structure and will take function pointers to the relevant vtable routines for data pass-through. We are going to decrease the number of PARROT_EXPORT routines from this subsystem and make a cleaner API.