Reworked timer handling to fix a race
If after an alarm expires a new alarm was set to the exact same time (now)
Parrot_alarm_runloop would not notice and still wait for new alarms without
notifying any threads. This happened quite often on Windows because
Parrot_floatval_time() gives only millisecond precision there which is an
eternity for a processor.

This reworked version is no longer dependent on any precision.