Fix socket tests, though I'm not happy with this fix.
If we do a recv for X bytes but only receive Y (Y < X), we shouldn't try to recv again to get more because recv can hang if there's no data incoming. Notice that this 'fix' fails for very large reads where we fill the buffer and could legitimately recv more bytes but do not.