Fix [GH #782] Installed parrot_config.o: Permission denied
make install may be done via sudo make install.
File::Copy::copy may drop the 644 permissions and leave only a
root owned 600 file of the files created as prep to the install
target. Which are installable's and install_config.o, which is
copied to parrot_config.o.

Use the newer File::Copy::cp instead, which works back to perl 5.6 (2.03),
but tries to keep the src perms starting with v2.14, with perl 5.10.1.