Release 4.6.0 - Wild Parrots of Telegraph Hill
See http://www.imdb.com/title/tt0424565/

"The first half of the documentary film focuses on Mark Bittner's
relationship with the Parrots living on the Telegraph Hill in the center of
San Francisco and how he came to be their caretakers.  He has a name for
each one and gives the audience stories of how he's interacted with them.
Included in these reminiscences, is general background information on bird
life in San Francisco and the various theories on how these South American
parrots came to exist in the Bay Area.

The second half of Parrots deals with the fate of the birds now that Mark
has to vacate his small home for the last three years.  Apparently, his care
of the birds has attracted world wide attention, especially in Europe.
There is never really any tension in the fate of Mark or the parrots since,
as Mark points out, the birds are perfectly capable of taking care of
themselves.  The viewer probably has a little less confidence in Mark
though, as he says he has no idea what he will do." - imdb review by
kasserine (US)