rename parrot-nqp to nqp
Also mark TODO portability in the makefile.
As discussed with pmichaud, we should git submodule to some safe-points,such as
nqp tags. jonathan is planning to revamp 6model a bit later, which should not affect us.
pmichaud does not want to abstract 6model into a seperate git repo, so we should
either copy it, or use it from ext/nqp-p6.