move os.pmc from dynpmc back to static pmc
Manually cleaned up pull request 778 by bdw.

Rationale whiteknight:
It's my belief that making OS PMC a dynpmc in the
first place was a mistake.  It is the only real way to access some pretty
fundamental filesystem and other functionality from the user level.
Further, all the real logic is implemented inside libparrot already, but
without OS PMC it isn't exposed to the user at all.

dynpmcs should be used for cases when we want to add new logic to the VM,
not when we already have that logic but want to be stingy about when we
expose it.

All our major users rely on OS pmc also. Winxed, Rakudo, NQP and Rosella all
make some use of it.  It's important and necessary, and it should be in
core.