[cage] Improve llvm detection
Probe for llvm-config-3.0 llvm-config-2.9 llvm-config-2.8 which do exist on debian,
where llvm-config does not exist.

Print found version, like "yes, 3.0" as with gcc and other libs.
If not requested, print "not requested" instead of "no" to give a hint,
that --with-llvm is wanted.

Add llvm and inline probes to ChangeLog