'1' added to result message is superfluous; revert.
The second argument passed to auto::llvm::_handle_result() is merely a Boolean
used to determine whether the result message displayed as part of
Configure.pl's output should be set to 'yes' or 'no'.  Consequently, a result
of 'yes, 1' conveys nothing above or beyond 'yes'.  Reverting the modification
(which was also causing failures in t/steps/auto/llvm-01.t).