Revert "Merge pull request #943 from gerdr/fix-socket-readline"
This reverts commit fb2e9570d25950c568edc50c1f9f5b1e0275bc2a, reversing
changes made to 83bb8639bc79699a16d594d53ca8abd92788ab03.

This is needed for the duration of the 5.2.0 release process, because the
branch was merged prematurely, and causes test failures. When the branch is
in a completed state, and ready to merge, it is important that a
"revert-the-revert" is performed just before the merge:
    git revert ID-of-*this*-revert
For rationale, see:
https://www.kernel.org/pub/software/scm/git/docs/howto/revert-a-faulty-merge.txt