[configure] Determine the cpu type and fill the appropriate config item
This information is needed in order to fill the PARROT_CPU_TYPE #define and
is required so that the CPU model (aka type) is specified as per the docs in
`src/dynoplibs/sys.ops`.

This currently only works on Linux; it *has* been tested, however not on
other platforms as yet.