Always use redirection and *NEVER* perldoc -d  ( github #520 )
perldoc -d $path is limited because perldoc SETUID's to `UID=nobody` if
`UID=0`, which basically guarantees that `perldoc -d $path` will not be
able to write anywhere in the working directory.

Though using redirection may be limited on some platforms, any competing
solution should change the code executed depending on platform.

On UNIX as UID=0, -d should **always** be avoided because of perldocs
inherent defects.