[config] support __builtin_expect LIKELY/UNLIKELY branch optimizations
add EXPECT,LIKELY,UNLIKELY macros to feature.h as in the linux kernel and perl5
probe for __builtin_expect() support in a new auto::expect configure step

Usage: if (UNLIKELY(var == 0)) fail();

Add emacs mode:c support for test_c.in files, and relax the coda test for this.
Should be used in all test_c.in files.