[imcc] rename to set_write_pasm, revamp api
Improve docs of the imcc API

pasm_out:
Pass the outfile path to set_write_pasm, so that we can
initialize it emit_open. Since the lexer doesn't know the outfile name
check for NULL in the lexer called emit_open(). Already initialized earlier
by the frontend.

The new parrot needs to store the outfile for the pasm out api.