runcores: fix docs, match docs with implementation and vice versa
use docs/running.pod as description for the runcore, debugging and
optimization options.
add missing docs for the runcores.

we already have a new parrot manpage for the rest.  allow the old runcores
in the implementation as described in the docs.
but disable the exec runcore, as it is different from the jit runcore.
there is now side-effect of producing a native executable (as of now).