[include] wrapping lines to 100 characters
This change is specific for PPC and thus I could not test it (Linux x86_64).
The change is trivial and *should* work, however it would be preferable for
someone with a PPC architecture to check that everything still works.  This
is the reason that this commit is separate to the others.

rurban: tested ok on PPC