[tools] correcting `rm_f` command in `make_cover.pl`
When removing `.gcov` files, the configured `rm_f` command was used.  This
unfortunately used an unexpanded `make` variable `$(PERL)` and thus was
producing errors such as:

    sh: 1: PERL: not found
    sh: 1: -MExtUtils::Command: not found

The fix was to use the configured perl value and the `ExtUtils::Command`
option as would be used within a Makefile.  This change means that what the
code intends to do (delete the `.gcov` file) actually happens and removes
the error from `make cover` output.