[pmc2c] support RETURN(PMC* const ...) and (INTVAL pos>>32)
const is no variable name.
allow full expressions as return value, not just names.

just skip ptr dereferences /^\**/. Note that the * return syntax is flawed,
as we should support both cases, return *ptr; and return ptr;
But we get away with it as we return only return (decl)*ptr and return (decl*)ptr;