[test] no gc-debug runcore when testing the fast runcore
With make fulltest and variants (t/fullharness) we omit -D040 and --gc-debug
when testing -f -O1 -O2 and the various set_usual_target() tests.
This caught some -O1/-O2 errors and speeds up the fulltest suite.
We only need to run a full GC after each op and randomize the N/I regs with the
normal internal testsuite, but not the extensions and various other tests.

See [GH #1086]

Current errors:
  testO1__t/op/calling.t, testO2__t/op/calling.t,
  testO2__t/compilers/imcc/syn/regressions.t