[io] optimize PIO_NR_OPEN from 256 to 3 [GH #1093]
PIO_NR_OPEN is now the number of statically allocated interpreter std
handles for stdin, stdout and stderr only.  It is only used safely in
one place, accessing only 0-2.
It has nothing to do with the total number of open filehandles, which
are allocated dynamically as PMC.
Leads to a small memory optimization, 253 less words for each interpreter.