[build] Enable --disable-shared, link to static libparrot
When the system did support dynamic linking we had no chance to
link against a static libparrot, as PARROT_LIBS checked for
has_dynamic_linking and not parrot_is_shared.
Note that we have no --disable-static yet, so we cannot forbid
creating a LIBPARROT_STATIC, this only relies on has_static_linking.