coro: refactor WIP, cheaper reset
do not delete the ctx for a reset. an empty ctx is for the case
calling a coro.
but autoreset is performed when returning from a coro.
so change the reset method to just turn back the adress to its
original, and handle autoreset with an inlined variant thereof.
But we still need to switch segments/ctx when returning from a
autoreset (returncc) coro.