[core] Downgrade external ascii on multi-byte platform encodings
We should scan those external strings for pure-ascii and set the
ascii encoding if the multibyte platform encoding is not needed.
This is mostly helpful to Getopt handling and is better then downgrade
to binary. We cannot set latin1 as this is incompatible to utf-8.
On Parrot_str_from_platform_cstring we can even reuse the setjmp for
the mult-byte exception in to_encoding.