[config] better icu probes, esp for --m overrides #1182
Add a new `has_libpath_override' config hash entry, with
lib or lib64, when the libpaths had to be adjusted from lib or lib64.
Do this for intel in init::defaults and for gcc in auto::gcc.
We do not more about the rest.

We cannot just blindly assume adding -Isystem will help finding the icu
headers returned by icu-config. We actually need to link the libs also
to detect possible cross-linking problems.
Provide a better error message on icu and --m failures.