[gc] fix SIGBUS in GC trace_mem_block on alignment-strict CPUs (armv5)
We traced all mem and stack pointers +1 and not by +4 or +8. This is not
only a huge performance loss, it is also illegal on certain RISC
architectures.  ptrdiff_t is not a pointer, it is an integer.
use size_t* instead.
Note that this is not repro in qemu, even not with echo 5>/proc/cpu/alignment
on armv5-linux.

Should fix GH #1200.