detect -Wcast-align violations [GH #1201]
We need an ARMv5 or MIPS platform which do require word aligned pointers
and throw these warnings.

TODO: We might need to add a config probe on those platforms for a
PARROT_HAS_STRICT_WORDALIGN ifdef and check the alignment then at
runtime with a possible needed memcpy before we can access the pointer
value.