[config] unquote the make clean args
We use now /bin/rm for $(RM_F) in the Makefile
which did not expand "$(INC_PMC_DIR)/pmc_*.h"
and many more generated files.
This led to leftovers, confusing make install.
E.g. the file include/pmc/pmc_default.h was not removed, and so was not rebuilt,
and so was not added to MANIFEST.generated, so t/distro/manifest_generated.t
failed.

`make clean realclean` does now the same as
`git clean -dxf`
Fixes GH #1194 and GH #1206.