Supported release 8.1.0 "Andean Parakeet"
+ Restore make configure_tests, remove outdated t/steps/auto/labs-01.t

The Andean parakeet (Bolborhynchus orbygnesius) is a small 160 mm (6 in)
parrot found in the upland wooded valleys of the central eastern Andes in
Bolivia and Peru, at elevations of 1,500 m (4,900 ft) to 5,000 m (16,400
ft).  It is solid green, darker above and lighter below, with some blue
visible on the wings when flying.  It has a pale bill and medium length
tail.  It is stockier and darker than the mountain parakeet.

Most flocks are small, but sometimes include over 300 individuals. The
Andean parakeet eats seeds, buds, and berries, and it breeds in burrows in
steep banks.

It was formerly known as B. andicolus - Wikipedia