install inst_libparrot.$VER.$SOEXT
over libparrot.$VER.$SOEXT, if it exists.
Fixes GH #1213
inst_libparrot has the final -install_name on darwin,
libparrot not.

broken with 7.11.0 and 8.1.0 for darwin binary builds only.
note that on darwin you are not allowed to use make -j>1 for
the installable target!

Also fixes the broken install_files --dry-run flag