[test] sub dispatch with builtin namespace
Subs in "built-in PMC" namespaces are treated as PMC methods even
without the :method attribute.

P1=new Integer
P1.foo()

namespace Integer
sub foo

Should error with
  Method 'foo' not found for invocant of class 'Integer'
as it happens with subs of new namespaces.

This allows overriding of builtin methods, but it also makes it
impossible to write "class methods" that translate behaviors, since the
presence or absence of the :method attribute on the sub is ignored.

See GH #304, prev. TT #1108