Use consistent syntax for import of File::Spec functions
Heretofore in Perl modules underneath config/, some File::Spec
functionality was imported as methods and some as functions.

  $ ack 'use File::Spec' config
  config/inter/libparrot.pm
  21:use File::Spec ();

  config/auto/perldoc.pm
  23:use File::Spec qw (catfile );

  config/auto/pmc.pm
  21:use File::Spec::Functions qw/catfile/;

  config/init/defaults.pm
  27:use File::Spec;

  config/init/hints.pm
  19:use File::Spec::Functions qw/catfile/;

  config/gen/config_pm.pm
  24:use File::Spec::Functions qw(catdir);

In the case of config/auto/perldoc, we use-d File::Spec but followed
that with an argument, as if we had intended to use
File::Spec::Functions.  In recent versions of Perl that gives a warning:

  Attempt to call undefined import method with arguments ("catfile") via
  package "File::Spec" (Perhaps you forgot to load the package?) at
  config/auto/perldoc.pm line 23.

.. which will eventually become an exception.

This commit standardizes the way we import and call such functionality.