[pmc] Support unicode strings for bitwise ops
bands, bors, bxors and bnots understand now more multi-byte strings.
Closes GH #848

Old docs:
Performs a bitwise C<OR> on two Parrot strings, performing type and encoding
conversions if necessary.  Returns the result as a new string.

Added now:
Same encodings are compared bitwise.  Different two-byte encodings are first
converted down to latin1, and if that fails upgraded to ucs2 and if that
fails to ucs4.
2 byte <=> 4 byte conversion is not supported, thus e.g. utf8->utf16 fails.
But 2 byte to 1 byte is attempted.